[Delta] ddt32c2s: Fix portpld deiver probe issue.
After adding the new adapter i2c-24, it will try to add a device with address 0x36 and
probe the device by parent device driver(sysfpga). It is because that the parent deivce
of i2c-24 is set to 3-0036, which is a client device instead of an i2c bus.

Set the parent device to an adapter(i2c bus, in this case is i2c-3) to solve this issue.